Disable Wredundant-decls warning
Arduino headers have several of these warnings, and I can't find a way
to suppress the warnings only for libraries. I tried using -isystem, but
that results in conflicting declaration errors unless the project's
src and include dirs also use -isystem, which is not an option since
that'd also disable warnings for first-party code.